Combine CardBlueprint and CardObject into one class Card
CardBlueprint and CardObject seems to play similar roles, the only differences are related to their attributes, so it is more optimal to have one class with optional attributes and properties to easily get their type
