--- magic_the_gathering/card_objects/non_permanent.py
@@ -1,5 +1,0 @@
-from magic_the_gathering.card_objects.base import CardObject
-
-
-class NonPermanentObject(CardObject):
-    pass
