--- magic_the_gathering/card_objects/token.py
@@ -1,5 +1,0 @@
-from magic_the_gathering.card_objects.permanent import PermanentObject
-
-
-class TokenObject(PermanentObject):
-    pass
